+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 
 ### Implemented enhancements
 - MVP-79 - Add react build tool
+- MVP-52 - Add login and sign up modals